Tying up various loose ends
Render tags on each topic view page.

Remove unsupported fields for first iteration

For the milestone 1.0 release, we are not going to support the following
properties:

- Event :: latitude and longitude
- Resource :: Image

Instead of prompting the user to specify the author of a topic, we will
just always set that to the user creating the entry until we find that
this needs to change.

Also fixed some styling and permission inconsistencies.

A date picker is provided for event start and end times