Pull License from MEAN branch
The original branch was under the MIT license. The Rails rewrite should, IMO, carry over the same license.

I have added the line `Copyright 2015 startupwichita.com contributors` to make it clear that everyone who contributes owns this code. This is the part I am unsure about and why I'm making this a PR instead of just pushing to master.

Ownership of the domain name, the server it's deployed on, the Startup Wichita branding etc are separate questions.