Remove turbolinks so select2 on tag entry works
It was failing to initialize because the JS that runs on those pages was not
being executed for the new field on the screen since it was loaded in with
turbolinks. I think the site performance is good enough that we can remove it
without much worry. This would have to be done in the future if we did any
sufficiently complex JS anyways.

If we *really* want to, we can add them back in later.