Close #24 - Filtering people index
The failure was due to the broken `unless` in the people index view. It was
assuming there were no results if the *non-featured* person collection was
empty. If the result only included `featured` people, it showed the "No people"
message.

Another issue was that if a person had the same tag on different attributes
(skills and roles, for example) they would be included twice. So we're filtering
out to only show unique results now.