Updates colors used in error explanation boxes
The colors that were being used in error explanation boxes didn't allow
the errors to be read because they were white text on a white
background.  This updates the error explanation boxes to use the site's
orange color for the border around the box, the site's cream color for
the background of the box and the site's charcoal color for the
explanation text.

Fixes #100